+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal.widgets 0.2.0.9000
 
+### Breaking change
+* Added `ignoreInit` argument to `verbatim_popup_srv`. By default this is set to `TRUE`. To use previous behavior set it to `FALSE`.
+
 # teal.widgets 0.2.0
 
 ### Breaking changes